refactor(services): extract data class models from oversized services
For each service that exposes a constellation of public data classes,
sealed types, and enums alongside the main class, those models move to
a sibling <ServiceName>Models.kt. The main service class (still the
single point of truth) shrinks accordingly.

- StorageAccessService.kt:    1841 -> 1781 (6 models)
- WebFictionService.kt:       1245 -> 1198 (4 models)
- PodcastService.kt:          1075 ->  945 (17 models)
- AdvancedMusicPlayerService.kt: 1024 -> 979 (4 models)
- CloudSyncService.kt:         957 ->  884 (7 models)
- GeminiService.kt:            890 ->  796 (9 models)  [now under guardrail]
- GeminiComicService.kt:       815 ->  677 (16 models) [now under]
- AIReadingProgressService.kt: 807 ->  667 (7 models)  [now under]

Follow-up work: StorageAccessService, WebFictionService, PodcastService,
AdvancedMusicPlayerService, CloudSyncService, MediaScannerService,
UniversalTagService, FanfictionEpubConversionService still hold a single
oversized class and need their methods extracted into collaborator files.
That decomposition is non-mechanical (state, lifecycle, DI graph) and is
deferred to a dedicated follow-up PR.
